# perf(swaps): optimize token processing selectors