--- a/app/reducers/swaps/swaps.test.ts
+++ b/app/reducers/swaps/swaps.test.ts
@@ -60,6 +60,7 @@ import reducer, {
   SWAPS_SET_HAS_ONBOARDED,
   swapsSmartTxFlagEnabled,
   swapsTokensObjectSelector,
+  swapsTokensMultiChainObjectSelector,
   selectSwapsChainFeatureFlags,
   getFeatureFlagChainId,
 } from './index';
@@ -289,9 +290,13 @@ describe('swaps reducer', () => {
       const liveState = reducer(initalState as any, action) as SwapsState;
 
       // Should set flags for ethereum (0x1)
-      expect(liveState['0x1'].featureFlags).toEqual(DEFAULT_FEATURE_FLAGS.ethereum);
+      expect(liveState['0x1'].featureFlags).toEqual(
+        DEFAULT_FEATURE_FLAGS.ethereum,
+      );
       // Should set flags for bsc (0x38)
-      expect((liveState['0x38'] as { featureFlags?: unknown }).featureFlags).toEqual(DEFAULT_FEATURE_FLAGS.bsc);
+      expect(
+        (liveState['0x38'] as { featureFlags?: unknown }).featureFlags,
+      ).toEqual(DEFAULT_FEATURE_FLAGS.bsc);
     });
 
     it('should handle testnet chain IDs in dev mode', () => {
@@ -308,9 +313,13 @@ describe('swaps reducer', () => {
         const liveState = reducer(initalState as any, action) as SwapsState;
 
         // Should use mainnet feature flags for testnet
-        expect(liveState['0x1'].featureFlags).toEqual(DEFAULT_FEATURE_FLAGS.ethereum);
+        expect(liveState['0x1'].featureFlags).toEqual(
+          DEFAULT_FEATURE_FLAGS.ethereum,
+        );
         // Should also set the testnet chain with mainnet flags
-        expect((liveState['0xaa36a7'] as { featureFlags?: unknown }).featureFlags).toEqual(DEFAULT_FEATURE_FLAGS.ethereum);
+        expect(
+          (liveState['0xaa36a7'] as { featureFlags?: unknown }).featureFlags,
+        ).toEqual(DEFAULT_FEATURE_FLAGS.ethereum);
       });
     });
 
@@ -335,8 +344,12 @@ describe('swaps reducer', () => {
       const newState = reducer(existingState as any, action) as SwapsState;
 
       // Should update feature flags but preserve other properties
-      expect(newState['0x1'].featureFlags).toEqual(DEFAULT_FEATURE_FLAGS.ethereum);
-      expect((newState['0x1'] as { someOtherProperty?: string }).someOtherProperty).toBe('value');
+      expect(newState['0x1'].featureFlags).toEqual(
+        DEFAULT_FEATURE_FLAGS.ethereum,
+      );
+      expect(
+        (newState['0x1'] as { someOtherProperty?: string }).someOtherProperty,
+      ).toBe('value');
     });
 
     it('should skip chains without valid chain IDs in name mapping', () => {
@@ -359,9 +372,13 @@ describe('swaps reducer', () => {
       const liveState = reducer(initalState as any, action) as SwapsState;
 
       // Should process valid chains
-      expect(liveState['0x1'].featureFlags).toEqual(DEFAULT_FEATURE_FLAGS.ethereum);
+      expect(liveState['0x1'].featureFlags).toEqual(
+        DEFAULT_FEATURE_FLAGS.ethereum,
+      );
       // Should not create entry for invalid chain
-      expect((liveState as Record<string, unknown>).invalidChain).toBeUndefined();
+      expect(
+        (liveState as Record<string, unknown>).invalidChain,
+      ).toBeUndefined();
     });
 
     it('should skip non-object feature flag values', () => {
@@ -583,8 +600,11 @@ describe('swaps reducer', () => {
       swaps: {
         featureFlags: globalFeatureFlags,
         ...Object.entries(chainFeatureFlags).reduce(
-          (acc, [chainId, flags]) => ({ ...acc, [chainId]: { featureFlags: flags } }),
-          {}
+          (acc, [chainId, flags]) => ({
+            ...acc,
+            [chainId]: { featureFlags: flags },
+          }),
+          {},
         ),
       },
     });
@@ -664,7 +684,7 @@ describe('swaps reducer', () => {
           },
         },
         chainFeatureFlags: {
-          '0x1': {},  // Empty feature flags
+          '0x1': {}, // Empty feature flags
         },
       });
 
@@ -678,7 +698,7 @@ describe('swaps reducer', () => {
 
     it('should return empty object when no chain entry exists', () => {
       const rootState = createTestState({
-        selectedChainId: '0x89',  // Chain ID not in swaps state
+        selectedChainId: '0x89', // Chain ID not in swaps state
         globalFeatureFlags: {
           smartTransactions: {
             globalSetting: true,
@@ -759,6 +779,400 @@ describe('swaps reducer', () => {
     });
   });
 
+  describe('swapsTokensMultiChainObjectSelector', () => {
+    beforeEach(() => {
+      jest.clearAllMocks();
+    });
+
+    it('returns combined tokens from all sources', () => {
+      // Given tokens exist in SwapsController and across multiple chains
+      jest.spyOn(tokensControllerSelectors, 'selectAllTokens').mockReturnValue({
+        '0x1': {
+          '0xUserAddress123': [
+            {
+              address: '0x0000000000000000000000000000000000000010',
+              symbol: 'USER-TOKEN1',
+              decimals: 18,
+              hasBalanceError: false,
+              image: 'user1.png',
+            },
+            {
+              address: '0x0000000000000000000000000000000000000011',
+              symbol: 'USER-TOKEN2',
+              decimals: 6,
+              hasBalanceError: false,
+              image: 'user2.png',
+            },
+          ],
+        },
+        '0x89': {
+          '0xUserAddress123': [
+            {
+              address: '0x0000000000000000000000000000000000000012',
+              symbol: 'POLYGON-TOKEN',
+              decimals: 18,
+              hasBalanceError: false,
+              image: 'polygon.png',
+            },
+          ],
+        },
+      });
+
+      const mockState = {
+        engine: {
+          backgroundState: {
+            SwapsController: {
+              tokens: [
+                {
+                  address: '0x0000000000000000000000000000000000000020',
+                  symbol: 'SWAPS-TOKEN1',
+                  decimals: 18,
+                  occurrences: 5,
+                  iconUrl: 'https://swaps1.url',
+                },
+                {
+                  address: '0x0000000000000000000000000000000000000021',
+                  symbol: 'SWAPS-TOKEN2',
+                  decimals: 6,
+                  occurrences: 3,
+                  iconUrl: 'https://swaps2.url',
+                },
+              ],
+            },
+            AccountsController: {
+              internalAccounts: {
+                selectedAccount: 'account-1',
+                accounts: {
+                  'account-1': {
+                    address: '0xUserAddress123',
+                  },
+                },
+              },
+            },
+          },
+        },
+      };
+
+      // When the selector is called
+      const result = swapsTokensMultiChainObjectSelector(mockState);
+
+      // Then it returns all unique token addresses with undefined values
+      expect(result).toStrictEqual({
+        '0x0000000000000000000000000000000000000020': undefined,
+        '0x0000000000000000000000000000000000000021': undefined,
+        '0x0000000000000000000000000000000000000010': undefined,
+        '0x0000000000000000000000000000000000000011': undefined,
+        '0x0000000000000000000000000000000000000012': undefined,
+      });
+    });
+
+    it('returns user tokens when SwapsController is empty', () => {
+      // Given no SwapsController tokens exist
+      jest.spyOn(tokensControllerSelectors, 'selectAllTokens').mockReturnValue({
+        '0x1': {
+          '0xUserAddress123': [
+            {
+              address: '0x0000000000000000000000000000000000000030',
+              symbol: 'ONLY-USER-TOKEN',
+              decimals: 18,
+              hasBalanceError: false,
+              image: 'onlyuser.png',
+            },
+          ],
+        },
+      });
+
+      const mockState = {
+        engine: {
+          backgroundState: {
+            SwapsController: {
+              tokens: [], // Empty SwapsController tokens
+            },
+            AccountsController: {
+              internalAccounts: {
+                selectedAccount: 'account-1',
+                accounts: {
+                  'account-1': {
+                    address: '0xUserAddress123',
+                  },
+                },
+              },
+            },
+          },
+        },
+      };
+
+      // When the selector is called
+      const result = swapsTokensMultiChainObjectSelector(mockState);
+
+      // Then it returns only the user tokens
+      expect(result).toStrictEqual({
+        '0x0000000000000000000000000000000000000030': undefined,
+      });
+    });
+
+    it('returns SwapsController tokens when user tokens are empty', () => {
+      // Given no user tokens exist across chains
+      jest
+        .spyOn(tokensControllerSelectors, 'selectAllTokens')
+        .mockReturnValue({});
+
+      const mockState = {
+        engine: {
+          backgroundState: {
+            SwapsController: {
+              tokens: [
+                {
+                  address: '0x0000000000000000000000000000000000000040',
+                  symbol: 'ONLY-SWAPS-TOKEN',
+                  decimals: 18,
+                  occurrences: 1,
+                  iconUrl: 'https://onlyswaps.url',
+                },
+              ],
+            },
+            AccountsController: {
+              internalAccounts: {
+                selectedAccount: 'account-1',
+                accounts: {
+                  'account-1': {
+                    address: '0xUserAddress123',
+                  },
+                },
+              },
+            },
+          },
+        },
+      };
+
+      // When the selector is called
+      const result = swapsTokensMultiChainObjectSelector(mockState);
+
+      // Then it returns only the SwapsController tokens
+      expect(result).toStrictEqual({
+        '0x0000000000000000000000000000000000000040': undefined,
+      });
+    });
+
+    it('removes duplicate tokens when same address exists in both sources', () => {
+      // Given the same token address exists in both SwapsController and user tokens
+      jest.spyOn(tokensControllerSelectors, 'selectAllTokens').mockReturnValue({
+        '0x1': {
+          '0xUserAddress123': [
+            {
+              address: '0x0000000000000000000000000000000000000050',
+              symbol: 'USER-VERSION',
+              decimals: 18,
+              hasBalanceError: false,
+              image: 'user.png',
+            },
+          ],
+        },
+      });
+
+      const mockState = {
+        engine: {
+          backgroundState: {
+            SwapsController: {
+              tokens: [
+                {
+                  address: '0x0000000000000000000000000000000000000050', // Same address
+                  symbol: 'SWAPS-VERSION',
+                  decimals: 18,
+                  occurrences: 10,
+                  iconUrl: 'https://swaps.url',
+                },
+              ],
+            },
+            AccountsController: {
+              internalAccounts: {
+                selectedAccount: 'account-1',
+                accounts: {
+                  'account-1': {
+                    address: '0xUserAddress123',
+                  },
+                },
+              },
+            },
+          },
+        },
+      };
+
+      // When the selector is called
+      const result = swapsTokensMultiChainObjectSelector(mockState);
+
+      // Then it returns only one entry for the duplicated address
+      expect(result).toStrictEqual({
+        '0x0000000000000000000000000000000000000050': undefined,
+      });
+      expect(Object.keys(result)).toHaveLength(1);
+    });
+
+    it('converts addresses to lowercase', () => {
+      // Given addresses with mixed case formatting
+      jest.spyOn(tokensControllerSelectors, 'selectAllTokens').mockReturnValue({
+        '0x1': {
+          '0xUserAddress123': [
+            {
+              address: '0X0000000000000000000000000000000000000060', // Uppercase
+              symbol: 'UPPER-TOKEN',
+              decimals: 18,
+              hasBalanceError: false,
+              image: 'upper.png',
+            },
+          ],
+        },
+      });
+
+      const mockState = {
+        engine: {
+          backgroundState: {
+            SwapsController: {
+              tokens: [
+                {
+                  address: '0X0000000000000000000000000000000000000070', // Uppercase
+                  symbol: 'SWAPS-UPPER',
+                  decimals: 18,
+                  occurrences: 1,
+                },
+              ],
+            },
+            AccountsController: {
+              internalAccounts: {
+                selectedAccount: 'account-1',
+                accounts: {
+                  'account-1': {
+                    address: '0xUserAddress123',
+                  },
+                },
+              },
+            },
+          },
+        },
+      };
+
+      // When the selector is called
+      const result = swapsTokensMultiChainObjectSelector(mockState);
+
+      // Then addresses are normalized to lowercase
+      expect(result).toStrictEqual({
+        '0x0000000000000000000000000000000000000070': undefined,
+        '0x0000000000000000000000000000000000000060': undefined,
+      });
+    });
+
+    it('returns empty object when no tokens exist', () => {
+      // Given no tokens exist in any source
+      jest
+        .spyOn(tokensControllerSelectors, 'selectAllTokens')
+        .mockReturnValue({});
+
+      const mockState = {
+        engine: {
+          backgroundState: {
+            SwapsController: {
+              tokens: [],
+            },
+            AccountsController: {
+              internalAccounts: {
+                selectedAccount: 'account-1',
+                accounts: {
+                  'account-1': {
+                    address: '0xUserAddress123',
+                  },
+                },
+              },
+            },
+          },
+        },
+      };
+
+      // When the selector is called
+      const result = swapsTokensMultiChainObjectSelector(mockState);
+
+      // Then it returns an empty object
+      expect(result).toStrictEqual({});
+    });
+
+    it('returns tokens from multiple blockchains', () => {
+      // Given tokens exist across Ethereum, Polygon, and BSC
+      jest.spyOn(tokensControllerSelectors, 'selectAllTokens').mockReturnValue({
+        '0x1': {
+          '0xUserAddress123': [
+            {
+              address: '0x0000000000000000000000000000000000000090',
+              symbol: 'ETH-TOKEN',
+              decimals: 18,
+              hasBalanceError: false,
+              image: 'eth.png',
+            },
+          ],
+        },
+        '0x89': {
+          '0xUserAddress123': [
+            {
+              address: '0x0000000000000000000000000000000000000091',
+              symbol: 'MATIC-TOKEN',
+              decimals: 18,
+              hasBalanceError: false,
+              image: 'matic.png',
+            },
+          ],
+        },
+        '0x38': {
+          '0xUserAddress123': [
+            {
+              address: '0x0000000000000000000000000000000000000092',
+              symbol: 'BSC-TOKEN',
+              decimals: 18,
+              hasBalanceError: false,
+              image: 'bsc.png',
+            },
+          ],
+        },
+      });
+
+      const mockState = {
+        engine: {
+          backgroundState: {
+            SwapsController: {
+              tokens: [
+                {
+                  address: '0x0000000000000000000000000000000000000080',
+                  symbol: 'SWAPS-TOKEN',
+                  decimals: 18,
+                  occurrences: 1,
+                },
+              ],
+            },
+            AccountsController: {
+              internalAccounts: {
+                selectedAccount: 'account-1',
+                accounts: {
+                  'account-1': {
+                    address: '0xUserAddress123',
+                  },
+                },
+              },
+            },
+          },
+        },
+      };
+
+      // When the selector is called
+      const result = swapsTokensMultiChainObjectSelector(mockState);
+
+      // Then it returns tokens from all chains plus SwapsController
+      expect(result).toStrictEqual({
+        '0x0000000000000000000000000000000000000080': undefined, // SwapsController
+        '0x0000000000000000000000000000000000000090': undefined, // Ethereum
+        '0x0000000000000000000000000000000000000091': undefined, // Polygon
+        '0x0000000000000000000000000000000000000092': undefined, // BSC
+      });
+      expect(Object.keys(result)).toHaveLength(4);
+    });
+  });
+
   it('should set has onboarded', () => {
     const initalState = reducer(undefined, emptyAction);
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -775,4 +1189,3 @@ describe('swaps reducer', () => {
     expect(liveState.hasOnboarded).toBe(true);
   });
 });
-